backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
pydicom
Pillow
numpy
python-dotenv
google-generativeai
inference-sdk
pyopenssl
pylibjpeg         
pylibjpeg-libjpeg  
pylibjpeg-openjpeg 
pytest
requests
pytest-mock